test: 단체 테스트 분리
