+++ src/test/java/kitchenpos/application/tablegroup/TableGroupServiceCreateTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.tablegroup;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,63 +9,16 @@ import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
-import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
-@ExtendWith(MockitoExtension.class)
-public class TableGroupServiceTest {
-
-    private List<OrderTable> standardOrderTables;
-    private TableGroup standardTableGroup;
-
-    @Mock
-    private OrderDao orderDao;
-
-    @Mock
-    private OrderTableDao orderTableDao;
-
-    @Mock
-    private TableGroupDao tableGroupDao;
-
-    @InjectMocks
-    private TableGroupService tableGroupService;
-
-    @BeforeEach
-    void setUp() {
-        OrderTable firstOrderTable = new OrderTable();
-        firstOrderTable.setId(1L);
-        firstOrderTable.setEmpty(true);
-        firstOrderTable.setNumberOfGuests(5);
-
-        OrderTable secondOrderTable = new OrderTable();
-        secondOrderTable.setId(2L);
-        secondOrderTable.setEmpty(true);
-        secondOrderTable.setNumberOfGuests(5);
-
-        standardOrderTables = new LinkedList<>();
-        standardOrderTables.add(firstOrderTable);
-        standardOrderTables.add(secondOrderTable);
-
-        standardTableGroup = new TableGroup();
-        standardTableGroup.setId(1L);
-        standardTableGroup.setCreatedDate(LocalDateTime.MAX);
-        standardTableGroup.setOrderTables(standardOrderTables);
-    }
+public class TableGroupServiceCreateTest extends TableGroupServiceTest {
 
     @DisplayName("단체를 생성할 때, 주문 테이블이 비어선 안 된다.")
     @Test
-    void createTableGroupWithNullOrderTables() {
+    void withNullOrderTables() {
         //given
         standardTableGroup.setOrderTables(null);
 
@@ -78,7 +31,7 @@ public class TableGroupServiceTest {
 
     @DisplayName("단체를 생성할 때, 주문 테이블이 2개보다 작아선 안 된다.")
     @Test
-    void createTableGroupWithBelowTwoOrderTables() {
+    void withBelowTwoOrderTables() {
         //given
         List<OrderTable> zeroOrderTables = new LinkedList<>();
         standardTableGroup.setOrderTables(zeroOrderTables);
@@ -92,7 +45,7 @@ public class TableGroupServiceTest {
 
     @DisplayName("단체를 생성할 때, 저장한 주문 테이블과 원래 테이블의 갯수가 달라선 안된다.")
     @Test
-    void createTableGroupWithChangedTableNumbers() {
+    void withChangedTableNumbers() {
         //given
         List<OrderTable> zeroOrderTables = new LinkedList<>();
         given(orderTableDao.findAllByIdIn(Arrays.asList(1L, 2L))).willReturn(zeroOrderTables);
@@ -106,7 +59,7 @@ public class TableGroupServiceTest {
 
     @DisplayName("단체를 생성할 때, 저장한 주문 테이블이 비어있어야만 한다.")
     @Test
-    void createTableGroupWithFullTable() {
+    void withFullTable() {
         //given
         standardTableGroup.getOrderTables()
             .forEach(table -> table.setEmpty(false));
@@ -121,7 +74,7 @@ public class TableGroupServiceTest {
 
     @DisplayName("단체를 생성할 때, 저장된 테이블들이 테이블 그룹에 할당되어 있으면 안된다.")
     @Test
-    void createTableGroupWithGrouppedTable() {
+    void withGrouppedTable() {
         //given
         given(orderTableDao.findAllByIdIn(Arrays.asList(1L, 2L))).willReturn(standardOrderTables);
         standardTableGroup.getOrderTables()
@@ -152,19 +105,4 @@ public class TableGroupServiceTest {
         );
     }
 
-    @DisplayName("단체를 삭제시, 요리하거나 식사중인 테이블이 없어야만 한다.")
-    @Test
-    void deleteGroupWithCookOrMeal() {
-        //given
-        given(orderTableDao.findAllByTableGroupId(1L)).willReturn(standardOrderTables);
-        given(orderDao.existsByOrderTableIdInAndOrderStatusIn(
-            Arrays.asList(1L, 2L),
-            Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))).willReturn(true);
-
-        //when
-
-        //then
-        assertThatThrownBy(() -> tableGroupService.ungroup(1L))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
 }
